Retrieve API Usage for specific methods

Get a summary of the API usage for your key.

Usage

get_api_usage(
  method = NULL,
  version = NULL,
  start_date = NULL,
  end_date = NULL
)

Arguments

method

(optional) a single API method for which you want data (e.g. GetDeparturesByStop)

version

(optional) a single API version for which you want data (e.g. 2.1)

start_date

(optional) the first day for which you want data, in YYYY-MM-DD format (cannot be earlier than 2013-02-07)

end_date

(optional) the last day for which you want data, in YYYY-MM-DD format (cannot be later than yesterday)

Value

date a date in YYYY-MM-DD format

calls the number of requests made to a method

versions a list of versions of the API that were used on a given date

number an API version number

methods a list of API methods that were used on a given date and within a given API version

name the name of an API method