Retrieve a List of CUMTD Bus Departures

Get a list of real-time departures for a specific stop_id.

Usage

get_departures_by_stop(stop_id, route_id = NULL, pt = NULL, count = NULL)

Arguments

stop_id

(required) id of the stop (ex. IT:1 or IT)

route_id

(optional) vector of route ids

pt

(optional) preview time in minutes between 0 and 60 (30 by default)

count

(optional) maximum number of departures you would like to recieve

Value

A data.frame with departure information for a specific stop:

destination trip's destination stop

expected expected departure time of the bus for the given stop

expected_mins number of minutes before expected departure time

headsign information usually shown on headsign

location latitude and longitude of vehicle

is_monitored whether the vehicle is communicating

is_scheduled whether this trip was scheduled

is_istop if this trip can be boarded without a fare/pass

origin trip's origin stop

route route information for the trip

scheduled scheduled departure time of the bus for the given stop

stop_id id of the stop the bus will be at

trip trip information for the departure

vehicle_id id associated with vehicle